basic-elt-pipeline-beam/pyproject.toml
[tool.poetry]
name = "etlpipelien"
version = "0.1.0"
description = ""
authors = ["sounish nath <flock.sinasini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
apache-beam = {extras = ["gcp"], version = "^2.56.0"}
dacite = "^1.8.1"
google-cloud-storage = "^2.16.0"
google-cloud-bigquery = "^3.22.0"
loguru = "^0.7.2"
setuptools = "^69.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
